Update 02_MSConvert-MALDI.en.md
Letting in that first scan is totally screwing up the analysis. The first scan is often garbage. I presumed MSconvert would be combining all the scans together, it's surprising 1 out of 119 scans is having such a big impact on the rest of the data if it is indeed averaging the data across the whole run. We are planning to set up some test acquisitions on the instrument to try and test what the software is doing.